+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ Version 0.11 (unreleased)
 
 * Optionally output to a user-specified array (``out`` keyword argument) when constructing
   geometries from ``indices`` (#380).
+* Added ``pygeos.empty`` to create a geometry array pre-filled with None or
+  with empty geometries (#381).
 
 **API Changes**